OpenHearing: TestSwitchedConnections: fix AIC errors being seen at startup
It turns out that we can't swtich AudioPaths until after the AICs have been enabled.  The problem is that the AudioPath setActive() method calls setupHardware(), which tries to interact with the AICs.  If the AICs haven't been enabled, you can't interact with them in the expected way.  So, the answer is simply to move the activateOneAudioPath() later in the setup() function, after the myTympan.enable() and earpieceShield.enable() have been called.